relocate all SetupSim tests under ./setupSim/... package
