+++ .circleci/config.yml
@@ -80,7 +80,24 @@ jobs:
             cd /tmp/go/src/github.com/FactomProject/factomd/
             go test $(glide nv | grep -v Utilities | grep -v longTest | grep -v peerTest | grep -v simTest | grep '...') -vet=off -v -cover -coverprofile=coverage.out
             goveralls -coverprofile=coverage.out -service=circle-ci -repotoken=$COVERALLS_TOKEN
-            bash -c "exit 0" # coverage should never fail
+
+  gofmt:
+    working_directory: /tmp
+    docker:
+      - image: circleci/golang:1.10
+
+    steps:
+      - attach_workspace:
+          at: /tmp
+
+      - run:
+          name: Gofmt test
+          no_output_timeout: 60
+          command: |
+            export PATH="/tmp/go/bin:$PATH"
+            export GOPATH=/tmp/go
+            cd /tmp/go/src/github.com/FactomProject/factomd/
+            ./test.sh gofmt
 
 
   # Docker builds
@@ -304,7 +321,7 @@ jobs:
 #      |
 #      ---------- test
 #                   |
-#                   ---------- coveralls
+#                   ---------- gofmt
 #                   |
 #                   ---------- docker_build
 #
@@ -324,7 +341,7 @@ workflows:
               only: /.*/
           requires:
             - build
-      - coveralls:
+      - gofmt:
           filters:
             tags:
               only: /.*/
@@ -350,7 +367,7 @@ workflows:
       - test:
           requires:
             - build
-      - coveralls:
+      - gofmt:
           requires:
             - test
       - docker_build:
