+++ activations/activationHeight.go
@@ -75,7 +75,7 @@ func (id ActivationType) String() string {
 
 	n, ok := ActivationNameMap[id]
 	if !ok {
-		n = fmt.Sprintf("ActivationId(%v)", id)
+		//n = fmt.Sprintf("ActivationId(%v)", id)
 	}
 	return n
 }
