--- engine/ci_whitelist
@@ -1,2 +1,0 @@
-engine/TestMakeALeader
-engine/TestAnElection
