--- simTest/ci_whitelist
@@ -1,4 +1,0 @@
-simTest/AddFNode_test.go
-simTest/EntriesBeforeChain_test.go
-simTest/HoldingRebound_test.go
-simTest/SetupANetwork_test.go
