+++ test.sh
@@ -1,82 +1,167 @@
 #!/usr/bin/env bash
 
-# this script is specified in .circleci/config.yml
-# to run as the 'tests' task
+# this script is specified as the 'tests' task in .circleci/config.yml
 
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )" # get dir containing this script
 cd $DIR # always from from script dir
 
+# set error on return if any part of a pipe command fails
+set -o pipefail
+
+# base go test command
+GO_TEST="go test -v -timeout=10m -vet=off"
+
+# list modules for CI testing
+function listModules() {
+  glide nv | grep -v Utilities | grep -v longTest | grep -v peerTest | grep -v simTest | grep -v activations | grep -v netTest | grep "\.\.\."
+}
+
+# formatted list of simTest/<testname>
+function listSimTest() {
+  go test --list=Test ./simTest/... | awk '/^Test/ { print "simTest/"$1 }'
+}
+
+# list of A/B Peer tests
+function listPeer() {
+  ls peerTest/*A_test.go
+}
+
+# load a list of tests to execute
+function loadTestList() {
+  case $1 in
+    unittest ) # run unit test in batches
+      TESTS=$({ \
+        listModules ; \
+      })
+      ;;
+    peertest ) # run only peer tests
+      TESTS=$({ \
+        ls peerTest/*A_test.go; \
+      })
+      ;;
+
+    simtest ) # run only simulation tests
+      TESTS=$({ \
+          listSimTest ; \
+      })
+      ;;
+
+    "" ) # no param: run everything
+
+      # running locally - default to all tests
+      if [[ "${CI}x" ==  "x" ]] ; then
+
+          TESTS=$({ \
+            listModules ; \
+            listSimTest ; \
+            listPeer ; \
+          })
+
+      else # running on circle
+
+        # run all tests 
+        if [[ "${CIRCLE_TAG}${CIRCLE_PULL_REQUEST}x" !=  "x" ]] ; then
+
+          TESTS=$({ \
+            listModules ; \
+            listSimTest ; \
+            listPeer ; \
+          } | circleci tests split ) # circleci helper spreads tests across containers
+
+        else # run single sim + all unit tests on every commit
+
+          TESTS=$({ \
+            listModules ; \
+            echo "simTest/TestAnElection" ; \
+          } | circleci tests split ) # circleci helper spreads tests across containers
+
+        fi
+      fi
+      ;;
+
+    * )
+      echo "Unknown option" $1
+      exit -1
+      ;;
+  esac
+}
+
+function testGoFmt() {
+  FILES=$( find . -name '*.go')
+
+  for FILE in ${FILES[*]} ; do
+    gofmt -w $FILE
+
+    if [[ $? != 0 ]] ;  then
+      FAIL=1
+      FAILURES+=($FILE)
+    fi
+  done
+
+}
+
 function runTests() {
-  if [[ "${CI}x" ==  "x" ]] ; then
-    # run locally
-    # 1. all unit tests except filtered packages
-    # 2. engine sim tests that are whitelisted
-    # 3. all files in simTest package
-    # 4. all sets of A/B tests in peerTest package
-    TESTS=$({ \
-      glide nv | grep -v Utilities | grep -v longTest | grep -v peerTest | grep -v simTest | grep -v elections | grep -v activations | grep -v netTest | grep "\.\.\." ; \
-      cat engine/ci_whitelist; \
-      ls simTest/*_test.go; \
-      ls peerTest/*A_test.go; \
-    })
-  else
-    # run on circle
-    # 1. run all unit tests
-    # 2. run only whitlisted tests in engine, peerTest, and simTest
-    TESTS=$({ \
-      glide nv | grep -v Utilities | grep -v longTest | grep -v peerTest | grep -v simTest | grep -v elections | grep -v activations | grep -v netTest | grep "\.\.\." ; \
-      cat */ci_whitelist; \
-    } | circleci tests split --split-by=timings)
-  fi
+  loadTestList $1
 
-	if [[ "${TESTS}x" ==  "x" ]] ; then
-    echo "No Tests"
-    exit 0
-  else
-    echo '---------------'
-    echo "${TESTS}"
-    echo '---------------'
-  fi
-
-  # NOTE: peer tests are expected to be named 
-  # in Follower/Network pairs
-  # Example:
-  #   BrainSwapA_test.go # 'A' runs first in background
-  #   BrainSwapB_test.go # 'B' test runs in foreground
-  BTEST="B_"
-  ATEST="A_"
-  FAILURES=()
-  FAIL=""
-
-  # exit code should fail if any part of the command fails
-  set -o pipefail
+  echo '---------------'
+  echo "${TESTS}"
+  echo '---------------'
 
   for TST in ${TESTS[*]} ; do
-    # start 'A' part of A/B test in background
-    if [[ `dirname ${TST}` == "peerTest" ]] ; then
-      ATEST_FILE=${TST/$BTEST/$ATEST}
-      TST=${TST/$ATEST/$BTEST}
-      echo "Concurrent Peer TEST: $ATEST_FILE"
-      nohup go test -v -timeout=10m -vet=off $ATEST_FILE &> testout.txt &
-    fi
+    case `dirname $TST` in
+      simTest )
+        testSim $TST
+        ;;
+      peerTest )
+        testPeer $TST
+        ;;
+      * ) # package name provided instead
+        unitTest $TST
+        ;;
+    esac
 
-    # run individual sim tests that have been whitelisted
-    if [[ `dirname ${TST}` == "engine" && ${TST/engine\//} != '...' ]] ; then
-      TST="./engine/... -run ${TST/engine\//}"
-      echo "Testing: $TST"
-    fi
-
-    echo "START: ${TST}"
-    echo '---------------'
-    go test -v -timeout=10m -vet=off $TST | tee -a testout.txt | egrep 'PASS|FAIL|RUN'
     if [[ $? != 0 ]] ;  then
       FAIL=1
       FAILURES+=($TST)
     fi
-    echo "END: ${TST}"
-    echo '---------------'
-
   done
+}
+
+# run A/B peer coodinated tests
+# $1 should be a path to a test file
+function testPeer() {
+  A=${1/B_/A_}
+  B=${1/A_/B_}
+
+  # run part A in background
+  nohup $GO_TEST $A &> a_testout.txt &
+
+  # run part B in foreground
+  $GO_TEST $B &> b_testout.txt
+}
+
+# run unit tests per module this ignores all simtests
+function unitTest() {
+  $GO_TEST $1 | grep -EHm1 "PASS|FAIL|panic|bind| Timeout "
+}
+
+# run a simtest
+# $1 matches simTest/<TestSomeTestName>
+function testSim() {
+  $GO_TEST -run=${1/simTest\//} ./simTest/... | grep -EHm1 "PASS|FAIL|panic|bind| Timeout "
+}
+
+function main() {
+  FAILURES=()
+  FAIL=""
+
+  if [[ "${1}" == "gofmt" ]] ; then
+    # check all go files pass gofmt
+    testGoFmt
+  else
+    # run tests
+    runTests $1
+  fi
 
   if [[ "${FAIL}x" != "x" ]] ; then
     echo "TESTS FAIL"
@@ -91,4 +176,4 @@ function runTests() {
   fi
 }
 
-runTests
+main $1
